requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.4.0
pillow>=8.0.0
